Add more color temperature choices to the "Preference" GUI.
And refactor code related to color temperature preferences.

I simplified the code that handles the case of upgrades to the
available choices. If we wanted to handle this properly we could add a
"flux version" key to the gconf settings, but I don't think it's worth
the extra complexity. Instead, we just change the users' color setting
to the default if their existing color gconf setting is no longer a
valid choice. Ultimately, I think it might be nice for the preference
menu to allow the user to enter an arbitrary number, in addition to
choosing from the defaults.

The main conceptual change in the refactor is to use a plain list of
color temps, instead of a dictionary mapping Glade/GTK menu indices to
color temps: the indices are implied by position in the list of
colors, so one less place for bugs.