Fix notorious `gtk-icon-theme-name` related warnings.
On startup we always see

    /usr/lib/python2.7/dist-packages/fluxgui/fluxapp.py:75: GtkWarning: Cannot transform xsetting Net/IconThemeName of type gchararray to type GdkColor

      gtk.gdk.screen_get_default().get_setting('gtk-icon-theme-name')
    /usr/lib/python2.7/dist-packages/fluxgui/fluxapp.py:75: GtkWarning: Cannot transform xsetting Net/IconThemeName of type gchararray to type gint

      gtk.gdk.screen_get_default().get_setting('gtk-icon-theme-name')

which users often include in bug reports (good on them!), but which is
ultimately irrelevant.

Turns out this was a lot more involved to fix than I expected. The fix
was to properly install the GTK theme specific icons in `setup.py`,
instead of using a runtime hack to choose the icon based on the current
theme name.

To install the icons properly I looked at `setup.py` from the [lookit
project](https://github.com/zachtib/lookit), which is where the hack I'm
removing came from (the lookit project removed that hack 5 years ago;
also caused spurious warnings for them). In short, instead of copying
three icons with different names to the "hicolor" theme dir, we copy a
separate icon with the same name to all three themes we care about,
"hicolor", "ubuntu-mono-dark", and "ubuntu-mono-light".