Fix #40: missing `xflux` binary on global manual install.
Turns the `xflux` binary was being installed with the wrong
permissions in the global, manual install, because it was installed as
"data" instead of as a "script". When installed as "data", its
permissions were `rwx------`. That's OK when installed locally, since
the owner is the user, but not OK when it's installed globally, since
the owner is root.