Document how to make global setup.py install work
This still includes the `gschema.xml` hacks, but solves the weird
permission problems. The key realization was that not all dirs created
by `setup.py` are listed in the `installed.txt`, but the permissions
of the created dirs is determined by the `umask`. This sure seems like
a bug in `setup.py`, but I haven't looked at their bug tracker yet.