setup.py: add posbility to disable compilation of gschema
When you install fluxgui, it also compiles the gschema, which is for
most distros fine. But on source distros like Gentoo, which do the
compilation of gschema separately, this will lead to a file collision,
as packages unter Gentoo are not allowed to install such files.

So we introduce a new env variable called DISABLE_GSCHEMA_COMPILED which
needs to be set to disable this behaviour.

Closes: https://github.com/xflux-gui/fluxgui/issues/124
Signed-off-by: Conrad Kostecki <conrad@kostecki.com>