Fix logic of lat+long settings check for Redshift
The Redshift backend requires that *both* the latitude and longitude
be set, but the logic in some places was only checking if *either* of
them were set, since that's what the analogous logic in the xflux
backend does (i.e. for xflux *either* the zip or the latitude must be
set).